requirements.txt
streamlit==1.21.0
rembg==2.0.23
Pillow==9.4.0
